Edit: remove duplicate helper file and reformat some tests
